+++ molo/core/templatetags/core_tags.py
@@ -7,7 +7,8 @@ from markdown import markdown
 
 from molo.core.models import (
     Page, ArticlePage, SectionPage, SiteSettings, Languages, Tag,
-    ArticlePageTags, SectionIndexPage)
+    ArticlePageTags, SectionIndexPage, ReactionQuestion,
+    ReactionQuestionChoice)
 
 register = template.Library()
 
@@ -81,7 +82,7 @@ def get_translation(context, page):
 def get_parent(context, page):
     parent = page.get_parent()
     if not parent.specific_class == SectionIndexPage:
-        return get_translation(context, parent)
+        return get_translation(context, parent.specific)
     return None
 
 
@@ -156,7 +157,7 @@ def topic_of_the_day(context):
 
 
 @register.inclusion_tag('core/tags/bannerpages.html', takes_context=True)
-def bannerpages(context):
+def bannerpages(context, position=None):
     request = context['request']
     locale = context.get('locale_code')
 
@@ -165,6 +166,18 @@ def bannerpages(context):
     else:
         pages = []
 
+    if position >= 0:
+        banners = get_pages(context, pages, locale)
+        if position > (len(banners) - 1):
+            return None
+        banners = get_pages(context, pages, locale)
+        if banners and len(banners) > position:
+            return {
+                'bannerpages': [banners[position]],
+                'request': context['request'],
+                'locale_code': locale,
+            }
+        return None
     return {
         'bannerpages': get_pages(context, pages, locale),
         'request': context['request'],
@@ -307,18 +320,101 @@ def load_child_articles_for_section(context, section, count=5):
     return articles
 
 
-@register.simple_tag(takes_context=True)
-def get_tag_articles(
-        context, latest_articles_count=12, section_count=1, tag_count=4,
-        sec_articles_count=4):
-    def get_positional_tag_articles(request, tag, exclude_list):
+def get_articles_for_tags_with_translations(
+        request, tag, exclude_list, locale, context, exclude_pks):
 
+    pks = [article_tag.page.pk for article_tag in
+           ArticlePageTags.objects.filter(tag=tag)]
+    return get_pages(
+        context, ArticlePage.objects.descendant_of(
+            request.site.root_page).filter(pk__in=pks).exclude(
+                pk__in=exclude_pks), locale)
+
+
+@register.assignment_tag(takes_context=True)
+def get_articles_for_tag(context, tag):
+    request = context['request']
+    locale = context.get('locale_code')
+    if tag:
+        main_tag = tag.specific.get_main_language_page()
         pks = [article_tag.page.pk for article_tag in
-               ArticlePageTags.objects.filter(tag=tag)]
+               ArticlePageTags.objects.filter(tag=main_tag)]
         return get_pages(
             context, ArticlePage.objects.descendant_of(
-                request.site.root_page).filter(pk__in=pks).exclude(
-                    pk__in=exclude_pks), locale)
+                request.site.root_page).filter(pk__in=pks), locale)
+    return None
+
+
+@register.assignment_tag(takes_context=True)
+def get_next_tag(context, tag):
+    request = context['request']
+    locale_code = context.get('locale_code')
+    current_tag = tag.get_main_language_page()
+    qs = Tag.objects.descendant_of(request.site.root_page).filter(
+        languages__language__is_main_language=True).live()
+    if qs:
+        tags = list(qs)
+        if len(tags) > 1:
+            if not (len(tags) == tags.index(current_tag) + 1):
+                next_tag = tags[tags.index(current_tag) + 1]
+            else:
+                next_tag = tags[0]
+
+            if next_tag.get_translation_for(
+                    locale_code, context['request'].site):
+                return next_tag.get_translation_for(
+                    locale_code, context['request'].site)
+            else:
+                return next_tag
+        else:
+            return None
+
+
+@register.assignment_tag(takes_context=True)
+def get_tags_for_section(context, section, tag_count=2, tag_article_count=4):
+    request = context['request']
+    locale = context.get('locale_code')
+    # Featured Tags
+    exclude_pks = []
+    tags_list = []
+    main_language_page = section.get_main_language_page()
+    child_articles = ArticlePage.objects.descendant_of(
+        main_language_page).filter(languages__language__is_main_language=True)
+    for article in child_articles:
+        exclude_pks.append(article.pk)
+    related_articles = ArticlePage.objects.filter(
+        related_sections__section__slug=main_language_page.slug)
+    for article in related_articles:
+        exclude_pks.append(article.pk)
+
+    tags = [
+        section_tag.tag.pk for section_tag in
+        section.get_main_language_page().specific.section_tags.all()
+        if section_tag.tag]
+    if tags and request.site:
+        qs = Tag.objects.descendant_of(
+            request.site.root_page).live().filter(pk__in=tags)
+        for tag in qs:
+            tag_articles = get_articles_for_tags_with_translations(
+                request, tag, exclude_pks, locale, context,
+                exclude_pks)
+            if len(tag_articles) > tag_article_count:
+                tag_articles = tag_articles[:tag_article_count]
+            exclude_pks += [p.pk for p in tag_articles]
+            tags = get_pages(context, qs.filter(pk=tag.pk), locale)
+            if tags:
+                tags_list.append((
+                    tags[0],
+                    tag_articles))
+    else:
+        return []
+
+    return tags_list
+
+
+@register.simple_tag(takes_context=True)
+def get_tag_articles(
+        context, section_count=1, tag_count=4, sec_articles_count=4):
 
     request = context['request']
     locale = context.get('locale_code')
@@ -328,16 +424,7 @@ def get_tag_articles(
     tags_list = []
     sections_list = []
 
-    # Latest Articles
-    latest_articles = request.site.root_page.specific.latest_articles(
-    )[:latest_articles_count]
-    exclude_pks += [p.pk for p in latest_articles]
-    data.update({
-        'latest_articles': get_pages(
-            context, ArticlePage.objects.filter(
-                pk__in=[p.pk for p in latest_articles]), locale)})
-
-    # Featured Section
+    # Featured Section/s
     sections = request.site.root_page.specific.sections()
     for section in sections[:section_count]:
         sec_articles = ArticlePage.objects.descendant_of(section).filter(
@@ -346,20 +433,51 @@ def get_tag_articles(
                 '-featured_in_homepage_start_date').exclude(
                 pk__in=exclude_pks)
         exclude_pks += [p.pk for p in sec_articles[:sec_articles_count]]
-        sections_list.append((
-            section,
-            get_pages(context, sec_articles, locale)[:sec_articles_count]))
+        section = SectionPage.objects.filter(pk=section.pk)
+        section_for_locale = get_pages(context, section, locale)
+        if section_for_locale:
+            section = section_for_locale[0]
+        sec_articles_for_locale = get_pages(context, sec_articles, locale)
+        if sec_articles_for_locale and len(
+                sec_articles_for_locale) > sec_articles_count:
+            sec_articles_for_locale = sec_articles_for_locale[
+                :sec_articles_count]
+        if section_for_locale:
+            sections_list.append((
+                section,
+                sec_articles_for_locale))
     data.update({'sections': sections_list})
 
-    # Featured Tags
-    for tag in Tag.objects.descendant_of(request.site.root_page).filter(
-            feature_in_homepage=True).live():
-        tag_articles = get_positional_tag_articles(
-            request, tag, exclude_pks)[:tag_count]
+    # Featured Tag/s
+    tag_qs = Tag.objects.descendant_of(request.site.root_page).filter(
+        feature_in_homepage=True).live()
+    for tag in tag_qs:
+        tag_articles = get_articles_for_tags_with_translations(
+            request, tag, exclude_pks, locale,
+            context, exclude_pks)
+        if tag_articles and len(tag_articles) > tag_count:
+            tag_articles = tag_articles[:tag_count]
         exclude_pks += [p.pk for p in tag_articles]
-        tags_list.append((tag, tag_articles))
-
+        tag_for_locale = get_pages(context, tag_qs.filter(pk=tag.pk), locale)
+        if tag_for_locale:
+            tags_list.append((
+                tag_for_locale[0],
+                tag_articles))
+        else:
+            tags_list.append((
+                tag,
+                tag_articles))
     data.update({'tags_list': tags_list})
+
+    # Latest Articles
+    data.update({
+        'latest_articles': get_pages(
+            context, ArticlePage.objects.descendant_of(
+                request.site.root_page).filter(
+                languages__language__is_main_language=True).exact_type(
+                    ArticlePage).exclude(pk__in=exclude_pks).order_by(
+                        '-featured_in_latest'), locale)})
+
     return data
 
 
@@ -369,7 +487,7 @@ def load_tags_for_article(context, article):
     request = context['request']
     tags = [
         article_tag.tag.pk for article_tag in
-        article.get_main_language_page().specific.nav_tags.all()
+        article.specific.get_main_language_page().nav_tags.all()
         if article_tag.tag]
     if tags and request.site:
         qs = Tag.objects.descendant_of(
@@ -377,6 +495,43 @@ def load_tags_for_article(context, article):
     else:
         return []
     return get_pages(context, qs, locale)
+
+
+@register.assignment_tag(takes_context=True)
+def load_choices_for_reaction_question(context, question):
+    locale = context.get('locale_code')
+    if question:
+        question_pk = question.get_main_language_page().pk
+        question = ReactionQuestion.objects.filter(pk=question_pk)
+    if question and question.first().get_children():
+        pks = [c.pk for c in question.first().get_children().filter(
+            languages__language__is_main_language=True)]
+        choices = ReactionQuestionChoice.objects.filter(pk__in=pks)
+        return get_pages(context, choices, locale)
+    return []
+
+
+@register.assignment_tag(takes_context=True)
+def load_reaction_question(context, article):
+    locale = context.get('locale_code')
+    request = context['request']
+    question = None
+    if article:
+        article_question = article.get_main_language_page() \
+            .reaction_questions.all().first()
+        if hasattr(article_question, 'reaction_question'):
+            question = article_question.reaction_question
+
+        if question and request.site:
+            qs = ReactionQuestion.objects.descendant_of(
+                request.site.root_page).live().filter(
+                    pk=question.pk, languages__language__is_main_language=True)
+        else:
+            return []
+        translated_question = get_pages(context, qs, locale)
+        if translated_question:
+            return get_pages(context, qs, locale)[0]
+        return question
 
 
 @register.assignment_tag(takes_context=True)