+++ molo/core/wagtail_hooks.py
@@ -1,6 +1,12 @@
 from django.conf.urls import url
 
+from molo.core.admin import ReactionQuestionsModelAdmin, \
+    ReactionQuestionsSummaryModelAdmin
+from molo.core.admin_views import ReactionQuestionResultsAdminView, \
+    ReactionQuestionSummaryAdminView
 from molo.core.models import LanguageRelation, PageTranslation, Languages
+from molo.core.utils import create_new_article_relations
+
 
 from django.core import urlresolvers
 from django.utils.translation import ugettext_lazy as _
@@ -11,7 +17,7 @@ from wagtail.wagtailcore.models import Page
 from wagtail.wagtailadmin.menu import MenuItem
 from wagtail.wagtailadmin.site_summary import SummaryItem
 from wagtail.wagtailadmin.widgets import ButtonWithDropdownFromHook
-
+from wagtail.contrib.modeladmin.options import modeladmin_register
 from wagtail.wagtailadmin.wagtail_hooks import page_listing_more_buttons
 
 from . import views
@@ -36,6 +42,30 @@ def register_admin_urls():
     ]
 
 
+@hooks.register('register_admin_urls')
+def register_question_results_admin_view_url():
+    return [
+        url(r'reactionquestion/(?P<parent>\d+)/results/$',
+            ReactionQuestionResultsAdminView.as_view(),
+            name='reaction-question-results-admin'),
+    ]
+
+
+modeladmin_register(ReactionQuestionsModelAdmin)
+
+
+@hooks.register('register_admin_urls')
+def register_article_question_results_admin_view_url():
+    return [
+        url(r'reactionquestion/(?P<article>\d+)/results/summary/$',
+            ReactionQuestionSummaryAdminView.as_view(),
+            name='reaction-question-article-results-admin'),
+    ]
+
+
+modeladmin_register(ReactionQuestionsSummaryModelAdmin)
+
+
 @hooks.register('construct_explorer_page_queryset')
 def show_main_language_only(parent_page, pages, request):
     main_language = Languages.for_site(request.site).languages.filter(
@@ -43,6 +73,12 @@ def show_main_language_only(parent_page, pages, request):
     if main_language and parent_page.depth > 2:
         return pages.filter(languages__language__locale=main_language.locale)
     return pages
+
+
+@hooks.register('after_copy_page')
+def add_new_tag_article_relations(request, page, new_page):
+    if new_page.depth < 3:
+        create_new_article_relations(page, new_page)
 
 
 @hooks.register('after_copy_page')